+++ fields/reveal/reveal.php
@@ -21,7 +21,16 @@ class RevealField extends BaseField {
 	public function element() {
 		$element = parent::element();
 		$element->data('field', self::$fieldname);
-		$element->data('refresh', c::get('plugin.reveal.refresh', 2000));
+
+		if(!site()->language()) {
+			$element->data('refresh', c::get('plugin.reveal.refresh', 2000));
+		} else {
+			if(site()->defaultLanguage()->code() != site()->language()->code()) {
+				$element->data('refresh', 0);
+			} else {
+				$element->data('refresh', c::get('plugin.reveal.refresh', 2000));
+			}
+		}
 		return $element;
 	}
 }